backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
yfinance==0.2.37
pandas==2.2.1
numpy==1.26.4
supabase==2.3.5
httpx==0.27.0
apscheduler==3.10.4
google-generativeai==0.5.4
python-dateutil==2.9.0.post0
scipy==1.12.0
requests==2.31.0
